+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- LWA support in the `amazon_creatorsapi.aio` API layer
+
+### Changed
+
+- Bumped `creatorsapi-python-sdk` from `1.1.2` to `1.2.0`
+- Updated bundled SDK support to include LWA endpoints for v3.x
+
 ## [6.1.0] - 2026-02-09
 
 ### Added